+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Map Changelog
 
 ## [Unreleased]
+### Changed
+- Configuration settings that take a list of files now support directories
+  as well.  The entire directory tree will be walked.
+
 ### Fixed
 - `Path.extension ()` properly finds the last dot in a path.